Fixed yield/ingredient number mismatch issue
Fixed: https://github.com/jonohein/PyOpenRecipe/issues/3
The amount and unit were in individual dictionaries for the yield. They need to be grouped together in the one dictionary per yield. This was only an issue when formatted as: 'amount: x, unit: y' and not when 'unit: amount'

Also set up testing for multiple recipes